New management command to dump data from database
Can be used in order to re-import data into another db.
Meant to be used for developers environment to work with realistic data.

For example, to dump all the entries for the model Idol:
> python manage.py dump_model_data_for_local_import Idol

Can work with multiple models and will generate multiple dump files + one global dump file.

The order of dump when working with models that have foreign keys to each other is important, as they may not exist and the script will not create them.

The script will attempt to create models when working with many_to_many fields that link to a model that is not linked to a MagiCollection + doesn't contain any foreign keys or many to many itself.

It's not recommended to dump users, accounts, collectible models, and any other model that's related to users content.

Example of full dump for Bandori Party's current database (in correct order):
> python manage.py dump_model_data_for_local_import Member Card Event Gacha Song Rerun Item Area Asset Costume DonationMonth Prize